% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudarpackage.R
\name{Rchol_multiple}
\alias{Rchol_multiple}
\title{Function \code{Rchol_multiple}}
\usage{
Rchol_multiple(array)
}
\value{
array of matrices in  with lower cholesky factor in lower triangle
}
\description{
cholesky factorize many matrices in parallel
}

