% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudarpackage.R
\name{Rconstruct_prec}
\alias{Rconstruct_prec}
\title{Function \code{Rconstruct_prec}}
\usage{
Rconstruct_prec(xtx, Mk, lambda, tau, K, V)
}
\arguments{
\item{xtx}{(scaled) data dim-by-dim precision matrix}

\item{Mk}{cluster occupancy}

\item{lambda}{prior precision}

\item{tau}{error precision}
}
\value{
array of matrices
}
\description{
get conditional precision matrix for cluster-specific parameters
}

