% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudarpackage.R
\name{Rmy_reduce}
\alias{Rmy_reduce}
\title{Function \code{Rmy_reduce}}
\usage{
Rmy_reduce(vec)
}
\arguments{
\item{vec}{vector of doubles}
}
\value{
Double sum of vector elements
}
\description{
reduce a vector on the device
}

