% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cudarpackage.R
\name{Rsummary}
\alias{Rsummary}
\title{Function \code{Rsummary}}
\usage{
Rsummary(all, key, num_clusts)
}
\arguments{
\item{all}{matrix of ints (data)}

\item{key}{vector of ints}

\item{num_clusts}{integer}
}
\value{
List a matrix with sums of rows by cluster, table of occupancy counts
}
\description{
summarize clusters
}

